app/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-cloud-aiplatform>=1.38.0